Refactor to align new code with existing code
